feat: add id search field for manga index
Lets clients fetch one manga's full row by id through the scrape-free
index RPC instead of get, which triggers a live re-scrape (and can throw)
whenever the manga's data is stale. Widened Number's parse from u16 to
i32 since manga ids can exceed 65535.

+++ src/util/search/field.rs
@@ -86,7 +86,8 @@ impl SearchField {
 
                 let number = captures.get(2).unwrap().as_str();
 
-                if let Ok(number) = number.parse::<u16>() {
+                // i32 (not u16) so this also covers manga.id, which can exceed 65535.
+                if let Ok(number) = number.parse::<i32>() {
                     return Ok(Expr::cust_with_values(format!("{ident} {compare} $1"), vec![number]));
                 } else {
                     return Err(Status::invalid_argument(format!("Expected number but got {value}")));
